move fetch to the services folder
